feat: :rocket: Self-executing functions
There may be cases where you need to create a function and run it on the fly. In Javascript it is very easy to create self-executing functions. Basically, we just have to wrap the anonymous function in question in parentheses (we don't need it to have a name, since we won't be saving it) and then execute it.